renaming Window -> WindowingSystem
